requirements.txt
aiogram==2.25.1
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
aiohttp==3.8.5
torch==2.0.1
torchvision==0.15.2
albumentations==1.3.0
scipy==1.11.1
requests==2.31.0